Basic Router Functionality
Functionality for basic navigation

Add a navigation bar component that shows the links to the application's pages. It is very primitive and will need to have added functionality for login/logout as well as the profile pages. Add a link option for "My Profile" because that will be needed for users to edit their information.
Also add a PrivateRoute.js. Its intended function is to handle sign in and to return the appropriate page when a user is signed in or not. The authorization does not exist right at the time of commit.