Move pages into new folder
Moved pages into new folder

The motivation is organization, the pages ought to be put into one place, where components are what goes into each page. Thus, the folder is called "pages." Rename "SignUp.js" to "Signup.js" for simplicity.

Added NavigationBar component. For each page, shows a list of links to the appropriate page. Added an error page for when a URL is incorrectly typed.