Working signup/login
Basic Login and signup implemented

The username feature is a work in progress. Will need to use Information.js to store it in a easily accessed global area. Modified loginprompts to no longer have submit types as refreshing the page does not work with logging in.
For future, will need to implement error messages for login/signup, as well as carrying over the login information, including username, to the new information.js