Update Options.js
Add rudimentary addWorkout and removeWorkout methods

The addWorkout accepts a Workout class object and a string representing the day. In our Firestore, a workout will be an object filled with days of the week ("monday", "tuesday", ...) , which are also objects that store two strings: name and type, where both are strings that are initially empty. This is a planned organization for the workouts field that the app stores, and the two methods operate under this data organization.