Working Profile Display
Successfully pulls the profile from the auth instance's profile Document which is named by the user ID. For instance, if a user with username "user" has UID AATlEL0ZIMNwFgFYUD9eMJOCUGd2, then the document that stores the all the user's information has name "AATlEL0ZIMNwFgFYUD9eMJOCUGd2."

UserData.js pulls the document and returns an object which can be read into a state variable as used in ProfileInformation.js. Please follow the code when looking to pull information from the database.