Fixed UpdateProfile Errors
Correctly implemented update profile feature

Before, the compiler would throw an error regarding the doc() call which was empty. Fix so that the props.id field informs the application whether to load in the updateProfile section, which should only happen for the profile page of the user and not of a different user resulting from a profile search.

Also, cleaned up a lot of imports that were unused and inputted informatory comments on updateProfile.js