Added comments, modified Options.js, fix login issue
Add pseudocode and modify Options.js

The pseudocode is to help anyone who wants to understand the code. In Options.js, modified methods to correctly update the correct workout fields in the server-side copy. Before the variables would be interpreted as a field itself rather than the variables' values (i.e. "workout_name" as a literal name for a field does not exist). Included a method to fetch the workouts field from the Firestore server in a very similar manner to the method in ProfileInformation.js. Cleaned up the organization of the code so as to get the document reference only once with the implementation of getDocument() stored into userRef (line 41 of Options.js). Modified methods to update the state variable workouts{} in addition to the server-side version.

Lastly, fixed an issue where logging out and logging in to a different account would not change the username displayed on Home.js