UserData.js Fix
Fix issue where getData() would incorrectly save the age to the favWorkout field.

Modified order of display in the Profile page so that Height comes first and ethnicity second. Done for consistency with rest of page and in back-end implementation, where all the code uses the same particular order for the fields.