Username Search
Implement search by username, returning a list of users

The search only looks for an exact match of username and so it will fail partial matches. Any users with the exact username typed in will be returned by the Firestore query. Specifically, the username's document will be returned. For now, render the username.