Profile Search Information Display
Implement method to display profile search results

The results of a profile search are rendered as a list of buttons, with the username of each matched user. Currently, all results will have the exact same name as the query because the search checks for strict username equality. Clicking on one of the buttons will render the profile information for that user. Only one person's information can be displayed at a time.

Also, implemented code so that when a user clicks "Update Profile," the user is sent back to the home page. This is done because the information does not show up on the profile page until it fetches from the database again. It is simple to force the user out of the profile page so that the update can register.