ProfileInformation Login Status Check
Fix so when no user is logged in, that the user goes to login

Previously, the application would crash because it would try to read from data, which is null. Instead, check before whether a user is logged in so that the user can login before trying to render nonexistent profile information.