Update ProfileSearch.js
Fix when user types own username

Previously, when user types their own username the search would return nothing. The intention was to not show the own user's information in the results but it went too far and returned no results. To fix this issue check for equality of User ID, which is much safer to check, as two users with the same username will not have the same user ID.