Implemented a general layout for the main page
Implemented the layout for the main page where the workouts and schedule are going to be. Used Grid from material-ui to do so and created a style file
that helps make the layout look right. Uses a temporary Item component to show where components are going to be laid out on the screen when inserted.
Header also has a profile button that has not been linked to the routes yet.