Implemented general structure for the main/home page
Moved the Layout.js code to Home.js Integrated that code with the existing code in Home.js. Linked the profile button in the Layout.js code so that it would route to the profile page. Home page
also welcomes user to site like it did before. Temporary placeholder components for showing how the page is structured is still in.