Fixing profile search bar and planner positioning
When loaded and rendered locally before the edit, the search bar appears behind the planner. Fixed it by adding styling to
the search bar that puts it above it. However, it might not be the best fix since the original author was using absolute
positioning for the css of these components, so it might differ from screen to screen. May have to properly fix later
time permitting.