Fetching and displaying the planner
Start work on displaying the planner

The style is messed up and the planner is in the wrong location. For source, code, there was a lot of reorganizing done to where certain components need to go. List.js holds Information.js and Options.js. Options.js is responsible for holding Planner.js.