Fixed display issues and restructured code so that it makes more sense
The display for the home page was broken after the last couple commits. Fixed that by restoring and modifying some code.
Took the new code from the past couple commits and made it so that the information for the workouts could be composed with the components more sensibly.
That also took care of some of the display issues.